Update style parser to support recursive rules
