+++ example/index.html
@@ -5,15 +5,18 @@
     <link rel="stylesheet" href="https://elect.in.th/assets/typography.css" />
 
     <script type="module">
+      import '../dist/staging-banner.js'
       import '../dist/social-sharer.js'
+      import '../dist/footer.js'
+      import '../dist/navigation-bar.js'
     </script>
   </head>
 
   <body>
+    <!-- <staging-banner /> -->
+    <!-- <navigation-bar /> -->
     <h1>Social Sharer Demo</h1>
-    <social-sharer />
-    <div style="background-color:black">
-      <social-sharer light="true" vertical="true" />
-    </div>
+    <!-- <social-sharer /> -->
+    <elect-footer />
   </body>
 </html>
